function/requirements.txt
numpy
joblib
azure-storage-blob
pandas